+++ src/optimize.h
@@ -169,9 +169,9 @@ namespace ALM_NS
 
         int run_elastic_net_crossvalidation(const std::string job_prefix,
                                             const int maxorder,
-                                            const int M,
-                                            const int M_test,
-                                            const int N_new,
+                                            const size_t M,
+                                            const size_t M_test,
+                                            const size_t N_new,
                                             std::vector<double> &amat_1D,
                                             std::vector<double> &bvec,
                                             const double fnorm,
@@ -182,8 +182,8 @@ namespace ALM_NS
                                             const int verbosity,
                                             std::vector<double> &param_out);
 
-        int run_elastic_net_optimization(const int M,
-                                         const int N_new,
+        int run_elastic_net_optimization(const size_t M,
+                                         const size_t N_new,
                                          std::vector<double> &amat_1D,
                                          std::vector<double> &bvec,
                                          const double fnorm,
